+++ sdk/metricsadvisor/azure-ai-metricsadvisor/tests/async_tests/test_data_feeds_aad_async.py
@@ -15,7 +15,7 @@ from azure.ai.metricsadvisor.models import (
     AzureTableDataFeedSource,
     AzureBlobDataFeedSource,
     AzureCosmosDBDataFeedSource,
-    HttpRequestDataFeedSource,
+    AzureLogAnalyticsDataFeedSource,
     DataFeedMetric,
     DataFeedDimension,
     DataFeedSchema,
@@ -31,7 +31,7 @@ from azure.ai.metricsadvisor.models import (
     MongoDBDataFeedSource,
     MySqlDataFeedSource,
     PostgreSqlDataFeedSource,
-    ElasticsearchDataFeedSource,
+    AzureEventHubsDataFeedSource,
 )
 from base_testcase_aad_async import TestMetricsAdvisorAdministrationClientBaseAsync
 
@@ -58,7 +58,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "SqlServer")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertIsNotNone(data_feed.source.query)
                 self.assertEqual(data_feed.granularity.granularity_type, "Daily")
                 self.assertEqual(data_feed.schema.metrics[0].name, "cost")
@@ -121,7 +120,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "SqlServer")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertIsNotNone(data_feed.source.query)
                 self.assertEqual(data_feed.granularity.granularity_type, "Daily")
                 self.assertEqual(data_feed.granularity.custom_granularity_value, None)
@@ -215,7 +213,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "SqlServer")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertIsNotNone(data_feed.source.query)
                 self.assertEqual(data_feed.granularity.granularity_type, "Custom")
                 self.assertEqual(data_feed.granularity.custom_granularity_value, 20)
@@ -290,7 +287,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "AzureTable")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.table, "adsample")
                 self.assertEqual(data_feed.source.query, "PartitionKey ge '@StartTime' and PartitionKey lt '@EndTime'")
             finally:
@@ -331,7 +327,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "AzureBlob")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.container, "adsample")
                 self.assertEqual(data_feed.source.blob_template, "%Y/%m/%d/%h/JsonFormatV2.json")
             finally:
@@ -373,91 +368,9 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "AzureCosmosDB")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.database, "adsample")
                 self.assertEqual(data_feed.source.collection_id, "adsample")
                 self.assertEqual(data_feed.source.sql_query, "'SELECT * FROM Items I where I.Timestamp >= @StartTime and I.Timestamp < @EndTime'")
-            finally:
-                await self.admin_client.delete_data_feed(data_feed.id)
-
-    @AzureTestCase.await_prepared_test
-    async def test_create_data_feed_with_http_request_get(self):
-        name = self.create_random_name("httprequestfeedgetasync")
-        async with self.admin_client:
-            try:
-                data_feed = await self.admin_client.create_data_feed(
-                    name=name,
-                    source=HttpRequestDataFeedSource(
-                        url=self.http_request_get_url,
-                        http_method="GET"
-                    ),
-                    granularity=DataFeedGranularity(
-                        granularity_type="Daily",
-                    ),
-                    schema=DataFeedSchema(
-                        metrics=[
-                            DataFeedMetric(name="cost"),
-                            DataFeedMetric(name="revenue")
-                        ],
-                        dimensions=[
-                            DataFeedDimension(name="category"),
-                            DataFeedDimension(name="city")
-                        ],
-                    ),
-                    ingestion_settings=DataFeedIngestionSettings(
-                        ingestion_begin_time=datetime.datetime(2019, 10, 1),
-                    ),
-
-                )
-
-                self.assertIsNotNone(data_feed.id)
-                self.assertIsNotNone(data_feed.created_time)
-                self.assertIsNotNone(data_feed.name)
-                self.assertEqual(data_feed.source.data_source_type, "HttpRequest")
-                self.assertIsNotNone(data_feed.source.url)
-                self.assertEqual(data_feed.source.http_method, "GET")
-
-            finally:
-                await self.admin_client.delete_data_feed(data_feed.id)
-
-    @AzureTestCase.await_prepared_test
-    async def test_create_data_feed_with_http_request_post(self):
-        name = self.create_random_name("httprequestfeedpostasync")
-        async with self.admin_client:
-            try:
-                data_feed = await self.admin_client.create_data_feed(
-                    name=name,
-                    source=HttpRequestDataFeedSource(
-                        url=self.http_request_post_url,
-                        http_method="POST",
-                        payload="{'startTime': '@StartTime'}"
-                    ),
-                    granularity=DataFeedGranularity(
-                        granularity_type="Daily",
-                    ),
-                    schema=DataFeedSchema(
-                        metrics=[
-                            DataFeedMetric(name="cost"),
-                            DataFeedMetric(name="revenue")
-                        ],
-                        dimensions=[
-                            DataFeedDimension(name="category"),
-                            DataFeedDimension(name="city")
-                        ],
-                    ),
-                    ingestion_settings=DataFeedIngestionSettings(
-                        ingestion_begin_time=datetime.datetime(2019, 10, 1),
-                    ),
-
-                )
-
-                self.assertIsNotNone(data_feed.id)
-                self.assertIsNotNone(data_feed.created_time)
-                self.assertIsNotNone(data_feed.name)
-                self.assertEqual(data_feed.source.data_source_type, "HttpRequest")
-                self.assertIsNotNone(data_feed.source.url)
-                self.assertEqual(data_feed.source.http_method, "POST")
-                self.assertEqual(data_feed.source.payload, "{'startTime': '@StartTime'}")
             finally:
                 await self.admin_client.delete_data_feed(data_feed.id)
 
@@ -501,7 +414,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "AzureApplicationInsights")
-                self.assertIsNotNone(data_feed.source.api_key)
                 self.assertEqual(data_feed.source.application_id, "3706fe8b-98f1-47c7-bf69-b73b6e53274d")
                 self.assertIsNotNone(data_feed.source.query)
 
@@ -544,7 +456,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "AzureDataExplorer")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.query, query)
 
             finally:
@@ -587,9 +498,7 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "InfluxDB")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertIsNotNone(data_feed.source.query)
-                self.assertIsNotNone(data_feed.source.password)
                 self.assertEqual(data_feed.source.database, "adsample")
                 self.assertEqual(data_feed.source.user_name, "adreadonly")
 
@@ -633,7 +542,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "AzureDataLakeStorageGen2")
-                self.assertIsNotNone(data_feed.source.account_key)
                 self.assertEqual(data_feed.source.account_name, "adsampledatalakegen2")
                 self.assertEqual(data_feed.source.file_system_name, "adsample")
                 self.assertEqual(data_feed.source.directory_template, "%Y/%m/%d")
@@ -677,7 +585,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "MongoDB")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.database, "adsample")
                 self.assertEqual(data_feed.source.command, '{"find": "adsample", "filter": { Timestamp: { $eq: @StartTime }} "batchSize": 2000,}')
 
@@ -718,7 +625,6 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "MySql")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.query, "'select * from adsample2 where Timestamp = @StartTime'")
 
             finally:
@@ -758,52 +664,7 @@ class TestMetricsAdvisorAdministrationClientAsync(TestMetricsAdvisorAdministrati
                 self.assertIsNotNone(data_feed.created_time)
                 self.assertIsNotNone(data_feed.name)
                 self.assertEqual(data_feed.source.data_source_type, "PostgreSql")
-                self.assertIsNotNone(data_feed.source.connection_string)
                 self.assertEqual(data_feed.source.query, "'select * from adsample2 where Timestamp = @StartTime'")
-
-            finally:
-                await self.admin_client.delete_data_feed(data_feed.id)
-
-    @AzureTestCase.await_prepared_test
-    async def test_create_data_feed_with_elasticsearch(self):
-        name = self.create_random_name("elasticasync")
-        async with self.admin_client:
-            try:
-                data_feed = await self.admin_client.create_data_feed(
-                    name=name,
-                    source=ElasticsearchDataFeedSource(
-                        host="ad-sample-es.westus2.cloudapp.azure.com",
-                        port="9200",
-                        auth_header=self.elasticsearch_auth_header,
-                        query="'select * from adsample where timestamp = @StartTime'"
-                    ),
-                    granularity=DataFeedGranularity(
-                        granularity_type="Daily",
-                    ),
-                    schema=DataFeedSchema(
-                        metrics=[
-                            DataFeedMetric(name="cost", display_name="Cost"),
-                            DataFeedMetric(name="revenue", display_name="Revenue")
-                        ],
-                        dimensions=[
-                            DataFeedDimension(name="category", display_name="Category"),
-                            DataFeedDimension(name="city", display_name="City")
-                        ],
-                    ),
-                    ingestion_settings=DataFeedIngestionSettings(
-                        ingestion_begin_time=datetime.datetime(2019, 1, 1),
-                    ),
-
-                )
-
-                self.assertIsNotNone(data_feed.id)
-                self.assertIsNotNone(data_feed.created_time)
-                self.assertIsNotNone(data_feed.name)
-                self.assertEqual(data_feed.source.data_source_type, "Elasticsearch")
-                self.assertIsNotNone(data_feed.source.auth_header)
-                self.assertEqual(data_feed.source.port, "9200")
-                self.assertEqual(data_feed.source.host, "ad-sample-es.westus2.cloudapp.azure.com")
-                self.assertEqual(data_feed.source.query, "'select * from adsample where timestamp = @StartTime'")
 
             finally:
                 await self.admin_client.delete_data_feed(data_feed.id)