+++ sdk/metricsadvisor/azure-ai-metricsadvisor/tests/test_data_feeds.py
@@ -15,7 +15,7 @@ from azure.ai.metricsadvisor.models import (
     AzureTableDataFeedSource,
     AzureBlobDataFeedSource,
     AzureCosmosDBDataFeedSource,
-    HttpRequestDataFeedSource,
+    AzureLogAnalyticsDataFeedSource,
     DataFeedMetric,
     DataFeedDimension,
     DataFeedSchema,
@@ -31,7 +31,7 @@ from azure.ai.metricsadvisor.models import (
     MongoDBDataFeedSource,
     MySqlDataFeedSource,
     PostgreSqlDataFeedSource,
-    ElasticsearchDataFeedSource,
+    AzureEventHubsDataFeedSource,
 )
 from base_testcase import TestMetricsAdvisorAdministrationClientBase
 
@@ -117,7 +117,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "SqlServer")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertIsNotNone(data_feed.source.query)
             self.assertEqual(data_feed.granularity.granularity_type, "Daily")
             self.assertEqual(data_feed.granularity.custom_granularity_value, None)
@@ -209,7 +208,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "SqlServer")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertIsNotNone(data_feed.source.query)
             self.assertEqual(data_feed.granularity.granularity_type, "Custom")
             self.assertEqual(data_feed.granularity.custom_granularity_value, 20)
@@ -282,7 +280,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "AzureTable")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.table, "adsample")
             self.assertEqual(data_feed.source.query, "PartitionKey ge '@StartTime' and PartitionKey lt '@EndTime'")
         finally:
@@ -321,7 +318,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "AzureBlob")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.container, "adsample")
             self.assertEqual(data_feed.source.blob_template, "%Y/%m/%d/%h/JsonFormatV2.json")
         finally:
@@ -361,86 +357,9 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "AzureCosmosDB")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.database, "adsample")
             self.assertEqual(data_feed.source.collection_id, "adsample")
             self.assertEqual(data_feed.source.sql_query, "'SELECT * FROM Items I where I.Timestamp >= @StartTime and I.Timestamp < @EndTime'")
-        finally:
-            self.admin_client.delete_data_feed(data_feed.id)
-
-    def test_create_data_feed_with_http_request_get(self):
-        name = self.create_random_name("httprequestfeedget")
-        try:
-            data_feed = self.admin_client.create_data_feed(
-                name=name,
-                source=HttpRequestDataFeedSource(
-                    url=self.http_request_get_url,
-                    http_method="GET"
-                ),
-                granularity=DataFeedGranularity(
-                    granularity_type="Daily",
-                ),
-                schema=DataFeedSchema(
-                    metrics=[
-                        DataFeedMetric(name="cost"),
-                        DataFeedMetric(name="revenue")
-                    ],
-                    dimensions=[
-                        DataFeedDimension(name="category"),
-                        DataFeedDimension(name="city")
-                    ],
-                ),
-                ingestion_settings=DataFeedIngestionSettings(
-                    ingestion_begin_time=datetime.datetime(2019, 10, 1),
-                ),
-
-            )
-
-            self.assertIsNotNone(data_feed.id)
-            self.assertIsNotNone(data_feed.created_time)
-            self.assertIsNotNone(data_feed.name)
-            self.assertEqual(data_feed.source.data_source_type, "HttpRequest")
-            self.assertIsNotNone(data_feed.source.url)
-            self.assertEqual(data_feed.source.http_method, "GET")
-        finally:
-            self.admin_client.delete_data_feed(data_feed.id)
-
-    def test_create_data_feed_with_http_request_post(self):
-        name = self.create_random_name("httprequestfeedpost")
-        try:
-            data_feed = self.admin_client.create_data_feed(
-                name=name,
-                source=HttpRequestDataFeedSource(
-                    url=self.http_request_post_url,
-                    http_method="POST",
-                    payload="{'startTime': '@StartTime'}"
-                ),
-                granularity=DataFeedGranularity(
-                    granularity_type="Daily",
-                ),
-                schema=DataFeedSchema(
-                    metrics=[
-                        DataFeedMetric(name="cost"),
-                        DataFeedMetric(name="revenue")
-                    ],
-                    dimensions=[
-                        DataFeedDimension(name="category"),
-                        DataFeedDimension(name="city")
-                    ],
-                ),
-                ingestion_settings=DataFeedIngestionSettings(
-                    ingestion_begin_time=datetime.datetime(2019, 10, 1),
-                ),
-
-            )
-
-            self.assertIsNotNone(data_feed.id)
-            self.assertIsNotNone(data_feed.created_time)
-            self.assertIsNotNone(data_feed.name)
-            self.assertEqual(data_feed.source.data_source_type, "HttpRequest")
-            self.assertIsNotNone(data_feed.source.url)
-            self.assertEqual(data_feed.source.http_method, "POST")
-            self.assertEqual(data_feed.source.payload, "{'startTime': '@StartTime'}")
         finally:
             self.admin_client.delete_data_feed(data_feed.id)
 
@@ -482,7 +401,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "AzureApplicationInsights")
-            self.assertIsNotNone(data_feed.source.api_key)
             self.assertEqual(data_feed.source.application_id, "3706fe8b-98f1-47c7-bf69-b73b6e53274d")
             self.assertIsNotNone(data_feed.source.query)
 
@@ -523,7 +441,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "AzureDataExplorer")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.query, query)
 
         finally:
@@ -564,9 +481,7 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "InfluxDB")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertIsNotNone(data_feed.source.query)
-            self.assertIsNotNone(data_feed.source.password)
             self.assertEqual(data_feed.source.database, "adsample")
             self.assertEqual(data_feed.source.user_name, "adreadonly")
 
@@ -608,7 +523,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "AzureDataLakeStorageGen2")
-            self.assertIsNotNone(data_feed.source.account_key)
             self.assertEqual(data_feed.source.account_name, "adsampledatalakegen2")
             self.assertEqual(data_feed.source.file_system_name, "adsample")
             self.assertEqual(data_feed.source.directory_template, "%Y/%m/%d")
@@ -650,7 +564,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "MongoDB")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.database, "adsample")
             self.assertEqual(data_feed.source.command, '{"find": "adsample", "filter": { Timestamp: { $eq: @StartTime }} "batchSize": 2000,}')
 
@@ -689,7 +602,6 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "MySql")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.query, "'select * from adsample2 where Timestamp = @StartTime'")
 
         finally:
@@ -727,50 +639,7 @@ class TestMetricsAdvisorAdministrationClient(TestMetricsAdvisorAdministrationCli
             self.assertIsNotNone(data_feed.created_time)
             self.assertIsNotNone(data_feed.name)
             self.assertEqual(data_feed.source.data_source_type, "PostgreSql")
-            self.assertIsNotNone(data_feed.source.connection_string)
             self.assertEqual(data_feed.source.query, "'select * from adsample2 where Timestamp = @StartTime'")
-
-        finally:
-            self.admin_client.delete_data_feed(data_feed.id)
-
-    def test_create_data_feed_with_elasticsearch(self):
-        name = self.create_random_name("elastic")
-        try:
-            data_feed = self.admin_client.create_data_feed(
-                name=name,
-                source=ElasticsearchDataFeedSource(
-                    host="ad-sample-es.westus2.cloudapp.azure.com",
-                    port="9200",
-                    auth_header=self.elasticsearch_auth_header,
-                    query="'select * from adsample where timestamp = @StartTime'"
-                ),
-                granularity=DataFeedGranularity(
-                    granularity_type="Daily",
-                ),
-                schema=DataFeedSchema(
-                    metrics=[
-                        DataFeedMetric(name="cost", display_name="Cost"),
-                        DataFeedMetric(name="revenue", display_name="Revenue")
-                    ],
-                    dimensions=[
-                        DataFeedDimension(name="category", display_name="Category"),
-                        DataFeedDimension(name="city", display_name="City")
-                    ],
-                ),
-                ingestion_settings=DataFeedIngestionSettings(
-                    ingestion_begin_time=datetime.datetime(2019, 1, 1),
-                ),
-
-            )
-
-            self.assertIsNotNone(data_feed.id)
-            self.assertIsNotNone(data_feed.created_time)
-            self.assertIsNotNone(data_feed.name)
-            self.assertEqual(data_feed.source.data_source_type, "Elasticsearch")
-            self.assertIsNotNone(data_feed.source.auth_header)
-            self.assertEqual(data_feed.source.port, "9200")
-            self.assertEqual(data_feed.source.host, "ad-sample-es.westus2.cloudapp.azure.com")
-            self.assertEqual(data_feed.source.query, "'select * from adsample where timestamp = @StartTime'")
 
         finally:
             self.admin_client.delete_data_feed(data_feed.id)